rename static to statics
