+++ core/src/main/kotlin/imgui/api/main.kt
@@ -32,10 +32,9 @@ import imgui.classes.IO
 import imgui.classes.Style
 import imgui.font.FontAtlas
 import imgui.internal.*
-import imgui.internal.classes.PoolIdx
 import imgui.internal.sections.ItemFlag
 import imgui.internal.sections.or
-import imgui.static.*
+import imgui.statics.*
 import org.lwjgl.system.Platform
 import imgui.WindowFlag as Wf
 import imgui.internal.sections.DrawListFlag as Dlf
