+++ core/src/main/kotlin/imgui/classes/Context.kt
@@ -15,8 +15,7 @@ import imgui.internal.DrawData
 import imgui.internal.classes.*
 import imgui.internal.hashStr
 import imgui.internal.sections.*
-import imgui.static.*
-import kool.free
+import imgui.statics.*
 import java.io.File
 import java.nio.ByteBuffer
 import java.util.*
